Refactor recent recipes and recipe cards
