+++ src/routes/create/+page.svelte
@@ -6,7 +6,7 @@
   import { createMarkdown, validateMarkdownTemplate } from '$lib/pharser';
   import { NDKEvent } from '@nostr-dev-kit/ndk';
   import type { recipeTagSimple } from '$lib/consts';
-  import FeedItem from '../../components/FeedItem.svelte';
+  import FeedItem from '../../components/RecipeCard.svelte';
   import { browser } from '$app/environment';
   import { goto } from '$app/navigation';
   import { nip19 } from 'nostr-tools';
